Replace SyncUsernameCommand with SyncUserCommand
